fixed word not despawning
As a quick fix, sam added a timeout before the collision layer on word was connected to time out. As a result, there was a bug where word could shoot through mobs, if player was close to them. I made it such that words will remove themselves after 1s no matter what and remove collision with the world