removed some useless code and made aiming of number weapon better
made it such that the number weapon aims with a bit of the velocity component so that it is not always lacking behind the mob it is targeting. However, this some makes some of the meteors spawn outside of the map.

I think this new issue is due to the knockback being quite ridiculous and the position for the meteor to be spawned is calculated at that instant